backend/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.1
PyMySQL>=1.1.1
cryptography>=46.0.7
python-dotenv>=1.0.1
PyJWT>=2.9.0
neo4j>=5.28.2
